Return log setup single volume saved to session
https://eaflood.atlassian.net/browse/WATER-4852
https://github.com/DEFRA/water-abstraction-system/pull/1768

During work on the return log setup journey it was noticed that user entered data wasn't being saved to the session correctly. The data being saved incorrectly was being saved as a string whereas we want it to be saved as a number. This PR is fixing one of the property's singleVolumeQuantity to be saved under the correct type.