Fix Match & Allocate result in 2PT annual bill run (#1782)
https://eaflood.atlassian.net/browse/WATER-4201

In [Handle TPT licence flags during supp process - pt1](https://github.com/DEFRA/water-abstraction-system/pull/1771), we updated the `MatchAndAllocateService` to return licences, rather than whether licences were found.

This information was used by the process services to indicate whether there was anything to bill. If `false` two-part tariff annual would know to set the bill run to empty.

But we changed it for two-part tariff supplementary to return the licences, so we would know which `licence_supplementary_years` records to assign to the bill run.

What we didn't do is update the two-part tariff annual `ProcessBillRunService` to handle this changed return value. It's a scenario that would never happen in production, but we still ensure we handle it and even have a (now failing!) acceptance test specifically for it.

This applies the change we should have done as part of PR #1771 !