Return log setup single volume saved to session (#1783)
https://eaflood.atlassian.net/browse/WATER-4852
#1768

During work on the return log setup journey, it was noticed that user-entered data wasn't being saved to the session correctly. The data being saved incorrectly was saved as a string whereas we want it to be saved as a number. This PR is fixing one of the property's singleVolumeQuantity to be saved under the correct type.