Update notify services response (#1781)
* Update notify services response

https://eaflood.atlassian.net/browse/WATER-4904

The legacy code uses the 'water.scheduled_notifications' schema to store the returns notifications. As part of the refactor to replace the existing legacy notifications code we need to align our implementation of notifications.

This change adds additional fields to the notify services responses.

We have added the 'statusText' which is used to determine the state of a notify notification. This will be 'CREATED' on a successful response.

The 'statusText' is available for an error but the legacy code does not use / store this. So we are following the existing pattern.

We have also added the 'plaintext' field. This is the raw string of the notification sent to the recipient.