Update notifications setup scheduled notification presenter (#1784)
https://eaflood.atlassian.net/browse/WATER-4778

There was a previous assumption that we could separate a 'notification' (something we send to notify) and a 'scheduled notification' (the thing we store).

This is not the case as we are required to save the 'licences' to 'water.scheduled_notifications'. This is not a big issue as the legacy code saves some of the Notify request/response properties (just in  slightly different shape).

To implement this change we have decided to use one presenter and have it create a 'scheduled notification'. This object includes everything we need to send a notification to Notify and save a 'scheduled notification'.

The only major difference is Notify expects an options object. We handle this in the change to the batch notifications service.