Add 'messageRef' to notifications setup scheduled notifications presenter
https://eaflood.atlassian.net/browse/WATER-4778

The legacy code has the concept of 'message_refs' found in 'water.scheduled_notification.message_ref'. This is used to group notifications by a unique type.

This column is used by the legacy code in multiple query's to retrieve notifications.

This changes adds a key map for these 'message_refs'.

As we have introduced the concept of 'both' for a 'contact_type' we use what we call the primary recipient for the
'message_ref'. When the licence is 'registered' we use the 'Primary user', when the licence is 'unregistered' we use
the 'Licence holder'. This is consistent with other areas of our codebase (they use the same 'template_id').