Add notifications setup scheduled notifications update notify
 https://eaflood.atlassian.net/browse/WATER-4778

When a scheduled notification has been sent to Notify we need to capture the Notify response.

The legacy code stores parts of the notify response in 'water.scheduled_notification'.

This change replicates Notify data used in the legacy codes .

When successful we will receive a notify id to check the future status of the notification with notify.

When there is an error we capture the 'raw' error and store in the scheduled notification.