Return logs check page - enter links https://eaflood.atlassian.net/browse/WATER-4870
This PR will update the location that the enter volumes/readings links point to on the "Check details" page prior to work commencing on  tickets https://eaflood.atlassian.net/browse/WATER-4870 and https://eaflood.atlassian.net/browse/WATER-4813

Initially this link was designed to point to a single page that would handle both readings and volumes. But we think it would be better to split these out into their own separate pages.