Update notifications setup batch notifications to persist
https://eaflood.atlassian.net/browse/WATER-4778

When a notification has been sent to Notify we need to store the Notify response with our 'scheduled notification'.

This change adds persisting of the 'scheduled notification' to the batch notification service.

Some of the data used for the payload is not required when persisting and has been removed after it has been used from the 'scheduled notification' object.

This change improves the test to check the data going through the batch service is saved correctly.