Update notifications setup batch notifications to batch by amount
https://eaflood.atlassian.net/browse/WATER-4778

The Notify service imposes a rate limit of 3,000 requests per minute and restricts the number of statuses returned per API call to a maximum of 250. Given these constraints, this update introduces a batch processing mechanism to handle the lowest of the two limitations, which in this case is the 250-message status retrieval limit from Notify.

Thus, the system will process a maximum of 250 recipients per batch.

To allow for flexibility and future adjustments, we’ve added a configurable setting that enables us to modify the batch size if necessary. This also aids in testing the solution with smaller batches for validation purposes.

Additionally, this approach helps us stay within the constraints of PostgreSQL’s batch insert limit. While PostgreSQL supports larger batches, the Notify API’s limitation on status checks (250 per call) remains the primary constraint. Therefore, the Notify limit is the driving factor for determining the batch size.factor.