Handle credit-only TPT tariff supplementary bills
https://eaflood.atlassian.net/browse/WATER-4952

In testing the new two-part tariff supplementary functionality (see WATER-4201 ) we found [a scenario where the licence wasn't being flagged](https://github.com/DEFRA/water-abstraction-system/pull/1794).

If a non-chargeable charge version is added, or a new charge version with the two-part tariff flag removed, the licence would not be flagged for TPT supplementary.

We're sorting the flagging situation, but then we realised it wouldn't do much. After the bill run has been setup, the first step is match and allocate. But that service relies on fetching the applicable two-part tariff charge versions for the licences in the bill run.

In this scenario, though the licence is flagged it wouldn't make it into the bill run because the charge version for the period in question is now not two-part tariff.

What the users expect to happen is that the licence won't't appear in the review step, because whatever water they extracted won't be charged (so doesn't need to be reviewed).

But they would expect to see a bill raised, which contains just a credit for the now void second-part charge.

This change updates the engine to cater for this scenario (without breaking it!)