Remove licences filters from assign bill run service
Previously, we'd let match and allocate do its work, and then return the licences it processed. These would be what we used to determine which licence supplementary year records to assign to the bill run.

But now we have to deal with licences that won't be processed by match and allocate.

So, we're switching the order of things when processing a two-part tariff supplementary bill run. Now, we'll assign _all_ matching licence supplementary year records to the bill run, at the time it is triggered.

We'll revert match & allocate back to its previous form (returning a boolean true or false if anything was populated).

This means we'll have something that ties those licences which no longer have two-part tariff charge versions to the bill run. They won't appear in review, but we can deal with them afterwards when we come to generate the bill run.