Update the GenerateTwoPartTariffBillRunService
There is now going to be enough difference in the process for annual and supplementary that it makes sense to separate them into their own services.

But to know which one to call, we still need to fetch the bill run first. Rather than do that in the controller, we figured we should still have a single service that it calls that does this.

Then we realised that service could also handle the validation of whether we can proceed, and the updating of the status to `processing`.

So, we have refactored `GenerateTwoPartTariffBillRunService` to handle those aspects. It then determines which engine to call.