Fix TPT match & allocate for licence end b4 charge (#1806)
https://eaflood.atlassian.net/browse/WATER-4957

Whilst testing two-part tariff supplementary (see WATER-4201 ), we've found a scenario that causes the match & allocate step to error.

For example, licence 01/123 has the following charge versions (assume everything TPT that needs to be ticked is ticked!)

- 2022-04-01 to 2023-10-05 (abs period 1 Apr to 31 Mar)
- 2023-10-06 onwards (abs period 1 Apr to 31 Mar)

If we created a TPT bill run (annual or supplementary, it doesn't matter) for 2023-2024, the match & allocate engine will process both charge versions.

Now imagine, the licence gets revoked on 2023-08-01. To match charge versions to returns we have to work out the 'charge-period'. This is essentially the intersection of the charge version and licence dates, with the billing period.

In our example, there is no charge period for that second charge version: the licence ended before the charge version starts. This is the scenario the TPT billing engine, specifically the match & allocate engine, isn't handling.

This change updates it to handle this.