Move existing fetch to be TPT annual only
Where we are intending to make the change is in the FetchBillingAccounts service. For annual we _only_ care about billing accounts linked to charge versions that were reviewed.

For supplementary, in order to include billing accounts linked to charge versions/licences that didn't get reviewed, for example, a licence that has been made non-chargeable, we can no longer select based on `review_licences`.

Rather than make the existing fetch try and work for both scenarios, we're creating a new one specifically for supplementary.