Fix return log download csv for meter readings (#1807)
* Fix return log download csv for meter readings

https://eaflood.atlassian.net/browse/WATER-4860

During testing of [Download return log as CSV](https://github.com/DEFRA/water-abstraction-system/pull/1736), an error was encountered when attempting to download csv's for return logs for meter readings. This was found to be because of the missing `endDate` value needed to apply readings against each return submission line.

This PR fixes the failing return log csv download.