Fix broken match & allocate for supplementary
Earlier, we moved when licence supplementary year records get assigned to the bill run to be _before_ match and allocate is run.

What we failed to spot at the time is that this broke the match & allocate engine for supplementary. The where exists clause we apply when supplementary is true specifically only fetched charge versions linked to licences with supplementary year records that had **not** been assigned!

Our assigning them before hand meant `FetchChargeVersions` when supplementary was true was never fetching anything.

Having spotted that was the problem, and knowing that we now also needed to get the bill run ID all the way to `FetchChargeVersionsService`, we realised it would make more sense to just pass the bill run instance itself all the way down.

Having done this, `FetchChargeVersionsService` now has all the information it needs to formulate the query and apply the correct where exists clause, including _only_ fetching those assigned to the bill run.