Update notification setup batch notifications service to have a delay (#1808)
* Update notification setup batch notifications service to have a delay

https://eaflood.atlassian.net/browse/WATER-4778

The Notify service imposes a rate limit of 3,000 requests per minute and restricts the number of statuses returned per API call to a maximum of 250. Given these constraints, this update introduces a batch processing delay mechanism to keep out request below the rate limit.

The system will process a maximum of 1506 notifications per minute.

To allow for flexibility and future adjustments, we’ve added a configurable setting that enables us to modify the batch delay if necessary. This also aids in testing the solution with shorter batch delays for validation purposes.

We are limiting ourselves to 1,500 requests per minute. Given that each status request returns up to 250 statuses, this translates to a maximum of 6 batches per minute (1,500 requests / 250 statuses per
request = 6 batches).

To ensure we stay within this 6-batch-per-minute constraint, we introduce a **10-second delay** between each batch. This delay is calculated as: 60 seconds (1 minute) / 6 batches = 10 seconds per batch.

This delay ensures we don’t exceed 1,500 requests per minute, leaving enough room for other parts of the system to make requests to the service.

The call to get the statues will count towards the rate limit. But to keep the explanation simple it has been ignored. In reality the total calls per minute will be 1506 (6 additional calls to get the statues for each batch).