Housekeeping - stub SendTransactions instead
We were stubbing the request to the charge module which is a few steps removed from the service we actually call.

For 2PT supplementary testing we wanted to avoid having to generate lots of transaction data, and could do so by stubbing SendTransactions directly.

It works so nice, and we want to keep things consistent, that we're updating the two-part tariff annual process billing period service to match.