Update Process billing period to handle changes
The only change we needed to make was to remove a guard clause that bailed processing the bill licences early if no generated transactions were found. Now we let it continue on to process supplementary transactions, which will go fetch all the previous transactions and work out that something needs crediting.

Even if it doesn't, the result is the same: nothing gets added to the bill.