Handle credit-only TPT tariff supplementary bills  (#1804)
https://eaflood.atlassian.net/browse/WATER-4952

In testing the new two-part tariff supplementary functionality (see [WATER-4201](https://eaflood.atlassian.net/browse/WATER-4201) ) we found [a scenario where the licence wasn't being flagged](https://github.com/DEFRA/water-abstraction-system/pull/1794).

If a non-chargeable charge version is added or a new charge version with the two-part tariff flag removed, the licence will not be flagged for TPT supplementary.

We're sorting the flagging situation, but then we realised it wouldn't do much. After the bill run has been set up, the first step is to match and allocate. But that service relies on fetching the applicable two-part tariff charge versions for the licences in the bill run.

In this scenario, though the licence is flagged, it wouldn't make it into the bill run because the charge version for the period in question is now not two-part tariff.

The users expect the licence to not appear in the review step because the water they extracted won't be charged (so it doesn't need to be reviewed).

But they would expect to see a bill raised, which contains just a credit for the now void second-part charge.

Because of these changes for supplementary, we started by splitting some of the logic we were reusing across two-part tariff bill runs. Fetch billing accounts was a good example of this. For two-part tariff annual, fetching based on what has a 'review' record is sufficient because only licences that have been reviewed will be part of the bill run.

But for two-part tariff, needing to credit previous transactions for licences that are no longer two-part tariff required, building a separate fetch billing accounts service.

We also needed to tweak how the process worked. For a two-part tariff annual, no matches or the user removing all licences in review are sufficient to mark the bill run as empty.

For two-part tariff supplementary, just because match & allocate found nothing or the user removed all the licences in review doesn't mean that there aren't licences that still need to be processed. So, we had to change what happens when those events occur based on the type of bill run being processed.

The most significant change, however, is non-chargeable. If the legacy code assigned a billing account to the non-chargeable charge version, our new fetch billing accounts service would have been sufficient. But it doesn't. So, somehow, we had to come up with a way to still include these licences, without scrapping our 'billing-account' based billing engine. Hence the addition of `FetchNonChargeableBillingAccountsService` (check out its documentation for more details of the problem it is solving).

So, _all these changes_ update our supplementary engine to cater for users adding a non-tpt or non-chargeable charge version to a licence that has previously had a second-part charge bill.

[WATER-4201]: https://eaflood.atlassian.net/browse/WATER-4201?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ