Make PRESROC flag setting clearer
We know we'll never get a PRESROC charge version that starts after the current financial year because the service won't allow one to be created that starts after 2022-04-01.

But without writing a bunch more documentation, someone new to this won't know that. Therefore, we make the intent clearer: if the charge version starts after the current financial year end, we leave as is.

Then if the scheme is 'alcs' set the flag and return the result.