Fix assigning bill run to TPT supplementary licences
https://eaflood.atlassian.net/browse/WATER-4952

In [Handle credit-only TPT tariff supplementary bills](https://github.com/DEFRA/water-abstraction-system/pull/1804) we had to update how licences are assigned to a two-part tariff supplementary bill run. Previously, we assigned whatever licences the match & allocate engine picked up. But when we realised this excluded licences where the change had made them no longer two-part tariff, we had to amend it.

We updated `AssignBillRunToLicencesService` to assign the licence supplementary year records _before_ the match and allocate step, and removed the 'licences' filter. What we failed to do include a filter that constrained the assignment to licences in the same region as the bill run! 🤦

This is causing two-part tariff supplementary to fall over, because it is trying to send transactions to the Charging Module API where the region does not match the bill runs.

This change updates `AssignBillRunToLicencesService` to include that filter.