Add Notification setup notify status presenter
https://eaflood.atlassian.net/browse/WATER-4961

The legacy code maps GOV.UK Notify statuses to an internal representation. It also stores the 'raw' status from GOV.UK Notify.

The legacy code has three types of notification statuses it shows the user:

- error
- pending
- sent

A 'sent' notification could be the result of multiple status updates from GOV.UK Notify.

This change adds the same logic used in the legacy code to map GOV.UK Notify statuses to the expected internal statuses.

This change also updates a previous assumption of a 'created' status in the system which is not used and should be 'pending'.

When we make the initial call to notify we do not receive a status, but we do receive a 'statusCode' (201) and
'statusText' ("CREATED"). This is inferred to mean the notifications is in the "created" state. This is the initial
'notifyStatus' value set for a 'scheduledNotification' (when no error has occurred).

A GOV.UK Notify status should only remain 'created' for a few seconds as stated in the documentation.