Update service
Using a raw query, we realised that knowing the bill run record has everything we need to determine what supplementary year records to flag without the need to pass in additional parameters.

Also, with the number of scenarios to cover and the use of a raw DB query, we felt safer updating the tests to talk to the DB rather than stubbing the queries.