Fix dupe results in TPT supp. Fetch Billing Acc.
https://eaflood.atlassian.net/browse/WATER-4952

In [Handle credit-only TPT tariff supplementary bills](https://github.com/DEFRA/water-abstraction-system/pull/1804), we had to change how we were fetching the billing accounts to be processed as bills in the bill run.

This was because we had to 'break' the direct link between the licences match & allocate found and processed, and those that go through the process billing period engine.

This is because match & allocate can only handle two-part tariff charge versions, but process billing period needs to handle non-two-part tariff charge versions. This is so it can determine if a credit is needed, for example, when a a licence has had a new charge version added that removes the two-part tariff agreement.

So, rather than have a single fetch billing account service used by the 2 two-part tariff bill run types, we created a specific service for each engine.

The one we added for supplementary, eschews selecting only those licences with `review_licences` records. Instead, it fetches all billing accounts linked to a licence with a 'licence supplementary year' (LSY) record that has been assigned to the bill run.

That appeared fine, until our QA team reported an error in UAT. We were able to track down to be with our new supplementary `FetchBillingAccountsService`. If there are duplicate LSYs, the it is return duplicate charge versions against each billing account.

How can you get duplicate LSY records I hear you ask?

Imagine you have made a change to a licence. This causes it to get flagged for two-part tariff supplementary and the first entry is made to LSYs.

You then create a two-part tariff supplementary bill run, so the LSY is assigned to that bill run. In review though, you spot you've made a mistake. But instead of removing it from review and then making the change, you make the change _then_ remove it.

The supplementary flagging engine will create a new YSL record when it sees the existing one for the licence is already assigned to the bill run. This is to ensure that the new changes are not missed.

But in our example, we now have two LSY records for the licence. When the next two-part tariff supplementary bill run is created, both will get assigned to the bill run. And this is how we end up with `FetchBillingAccountsService` pulling duplicate charge version records.

Thankfully, a one word change is all that is needed to avoid this issue!