Add notification setup confirmation page (#1816)
* Add notification setup confirmation page

https://eaflood.atlassian.net/browse/WATER-4778

Previous work has been done to add the notification batch service that will send the notification to GOV.UK Notify and save the initial state to the database.

This change adds the confirmation pages controller, service and presenter.

When the user presses send on the notification setup check page they will be redirected to this confirmation page.

This concludes the user journey.