Update notification setup default event status
https://eaflood.atlassian.net/browse/WATER-4778

There was a previous assumption of using setting the event status to 'started' when it is created. This resulted in the event not showing on the legacy on the 'notifications/report' page.

This change updates the default status to 'completed'. This will not change in our code base.

This lead to exploring where the 'sent' and 'error' values on the options object are used in the legacy code. They are not used as far as we can tell.

The recipients count displayed on the 'notifications/report' page is taken from 'metadata.recipients'.

The errors are 'counted' based on a query in the legacy code and does not use the 'metadata.error' key.