Update notification setup submit check tests (#1823)
https://eaflood.atlassian.net/browse/WATER-4778

The previous test stubbed the 'CreateEventService', this was previously acceptable however due to a recent refactor where we return the event id from the 'SubmitCheckService'. We need to improve the assertion of the event id, by checking the event is created in the database.

This change aligns the tests with our other test pattern of using the real database data.

This change also adds a test for the 'BatchNotificationsService' this is a complicated service which is crucial to the whole notification process. This has been tested independent, however we want to assure that the function is called (and the args are correct) and is able to process as it falls outside of the async/await functionality.