Update notification setup notify update status (#1825)
https://eaflood.atlassian.net/browse/WATER-4778

A previous assumption was made that a 'scheduled_notification' could have the status
'pending'. This fails validation when rendering a report. The legacy UI has a mapper that converts a 'draft' or 'sending' status into a 'pending' status for the notifications/report{eventId} page.

As we do not have a 'draft' state we will use the 'sending' status.

Therefore, when we initially send our calls to Notify (and the notification does not error) we will set the 'scheduled_notification.status' to 'sending'.