Fix change to billing account plus all previous!
https://eaflood.atlassian.net/browse/WATER-4952

The specific scenario that led us to look again at the two-part tariff supplementary billing engine was a licence where

- a bill was raised in the 2023/24 two-part tariff annual bill run
- a new charge version is added
- the new charge version is for a different billing account

When the two-part tariff supplementary was run, a credit was not raised. But when we ran the standard supplementary, it was. This forced us to dig back into the standard supplementary engine. We know it starts with fetching the charge versions, and then has a complex means of translating these records into blank 'bills' which are then processed by its `ProcessBillingPeriodService`.

All our new engines have started with fetching billing accounts to avoid that overly-complex step, and it was this difference which led us to believe we needed the `FetchNonChargeableBillingAccountsService` to ensure charge versions that had been made non-chargeable would still be included in the bill run, so we could find the previous transactions from when they weren't non-chargeable.

Many [thousands of lines of code and a weekend](https://github.com/DEFRA/water-abstraction-system/pull/1804) went into that change!

But after looking at what standard supplementary does, and with a wiser head we were able to spot what the _actual_ key difference is. The standard supplementary fetch does not have a filter on when a charge version ends, only when it starts. This means it fetches _all_ charge versions for a licence that start before the billing period ends. This is how it ensures it picks up all previous transactions and correctly determines what needs crediting, or not.

If we do the same with `FetchBillingAccountsService`, it not only handles the 'change in billing account' scenario, but also all our non-chargeable/non-two-part tariff scenarios.

So, in this change we strip out all the changes we made to add `FetchNonChargeableBillingAccountsService` as it is redundant. We then also strip the following clause from `FetchBillingAccountsService`.

```javascript
    .where((builder) => {
      builder.whereNull('chargeVersions.endDate').orWhere('chargeVersions.endDate', '>=', billingPeriod.startDate)
    })
```