Update notification setup to facilitate 'ad-hoc' journey
https://eaflood.atlassian.net/browse/WATER-4962

The 'ad-hoc' journey does not exist in the legacy code.

This change updates the notification setup journey to send notifications using the 'ad-hoc' journey. The services for a user to send 'ad-hoc' licences have been added previously. But the focus was on getting the notifications to Notify and storing in 'water.scheduled_notifications' for the other journeys as they are based on the legacy code.

The 'ad-hoc' journey introduces a new 'subType' and 'messageRef' that conforms to the existing patterns. This allows the 'notifications/report' and 'notifications/report/{eventId}' pages to function as expected and render the 'ad-hoc' notification report.