Update notifications setup submit ad-hoc licence
https://eaflood.atlassian.net/browse/WATER-4962

The 'ad-hoc' journey does not exist in the legacy code.

This change updates the notification setup submit ad-hoc- licence service to provide a 'determinedReturnsPeriod' in the session.

This is required to allow the 'ad-hoc' journey to use the existing 'reminders' GOV.UK Notify templates to send the notifications.

A previous change has implemented the ability to send notifications for ad-hoc licences -  https://github.com/DEFRA/water-abstraction-system/pull/1828.