Add new tpt supplementary seeder
Before we strip out the work done to return non-chargeable, we want to ensure FetchBillingAccountsService will return all the data we need for the various scenarios we have been verifying so far.

That's a lot of data that needs setting up, so rather than add even more to the existing unit tests, we create a new seeder instead.