Return "not due yet" status
https://eaflood.atlassian.net/browse/WATER-4954

There is a need for a status "Not due yet" for a return that is visible to the user so it is clear to them which return is due for submission. This status should be shown for those returns (Quarterly and Annual) that are not due. This PR is adding that status to the view returns page.