Add Notify statuses service (#1830)
* Add Notify statuses service

https://eaflood.atlassian.net/browse/WATER-4904

This change adds the 'NotifyStatusesService' which checks the status of a multiple notifications in notify.

To retrieve the notifications specific to a 'group' a unique 'reference' is used in notify. We send a unique 'reference' for most notifications.

We can use this unique 'reference' with the Notify api to get multiple statuses.

There is a limit of 250 notifications returned. This can be 'paginated' in a way by providing the last notification id from the previous call.

Notify returns the most recent statuses initially - https://docs.notifications.service.gov.uk/node.html#olderthan-optional.

Unlike our previous Notify services it's tests cannot be used against the live GOV.UK Notify. They would be too complicated and would require using our code to set up tests (what if that code is broken).