Fix volumes on check page
Realised the issue is how the setup journey treats quantity on a line differently to how the submission does.

The submission _always_ treats quantity as the value in cubic metres. The setup journey on the overhand treats it as whatever the user entered.

So, if during setup the user selected litres as the unit of measure and entered 1,000 you would get

- setup line: quantity = 1000
- submission line: quantity = 1

So, for the edit journey, we have to take the submission line quantity and convert it to match what the setup journey expects, and store that in the session.