Fix 'return received' backlink for submit return
https://eaflood.atlassian.net/browse/WATER-4844

We've spotted that when you go to submit a return for the first time, the `< Back` takes you to the licence's returns tab, and not the 'view return log' page you've just come from.

> This is assuming you've not yet reached the `/check` page, at which point `< Back` will take you there.

This small change fixes it to return you to the return log you've just come from.