Add `marked` npm package for markdown rendering
As part of the work to migrate the legacy [view licence communications page](https://github.com/DEFRA/water-abstraction-system/pull/1810), we needed a method to
convert the `plaintext` provided as the template for the [Notify](https://www.notifications.service.gov.uk/) service. The [marked](https://www.npmjs.com/package/marked) npm package
was being used by the legacy team to convert the `plaintext` Notify template into [markdown](https://www.markdownguide.org/) for display on the legacy licence communications view. We are adding this package and utilising
it as a `Nunjucks filter` to convert the `plaintext` into markdown.

This PR adds the `marked` npm package to water abstraction system.