Fix 'return received' backlink for submit return (#1833)
https://eaflood.atlassian.net/browse/WATER-4844

We've spotted that when you submit a return for the first time, the `< Back` on the "When was the return received?" page takes you to the licence's returns tab, not the 'view return log' page you've just come from.

> This is assuming you've not yet reached the `/check` page, at which point `< Back` will take you there.

This small change fixes it to return you to the return log you've just come from.