Add job notifications status update service logic
https://eaflood.atlassian.net/browse/WATER-4961

As part of the work to migrate the legacy notifications into the system repo, we need to have a mechanism to check and update the notifications statuses.

This change updates the ProcessNotificationsStatusUpdatesService to get the events (which statuses needed to be updated), call notify to get a status update and then persists the updated status.

When an error occurs the error count will be updated for the event.