Add job notification fetch event notifications service
https://eaflood.atlassian.net/browse/WATER-4961

When checking the status of a notification with GOV.UK Notify we want to only check the status when a notification is in the 'sending' status.

This change adds the logic to fetch all 'scheduledNotification' (where the status is 'sending') for every 'event' with a status of 'completed'.

The legacy code may leave many events with statuses like 'processing' or 'processed' and scheduled notifications with
a 'draft' status lingering in the database. This fetch focuses on events that are considered completed and relevant
for notification reporting purposes.